Remove credentials.json
Even though the 'json' file is excluded in the .gitignore,
it is still being tracked because its part of the repo.
git rm --cached (filename)
stops it from being tracked without removing it from the file system.
I put a docstring in views.py explaining how the .json
file should be structured.