Timeframes - data retrieval with conditional times
The challenge this commit attempts to solve is if the server/
database connection hasn't been running for a long time.
We'd then be missing data from the database and will have to
retrieve the data that's missing from that time frame.

To accomplish this, we'll have to try and check the most recent item
in the database, and if its longer than 15 minutes old, retrieve
the data from the last time there was data.