Celery Task works with database
The celery task successfully pulls data from the database.
This is critical because Celery is going to be taking
over the 'populate' view and needs to be able to manipulate
the database.

Next steps are to use BLEAT to run this on a schedule.

Another thing that needs to be figured out is maybe how to combine all
these three commands. Redis, celery worker and python runserver all need
to be running for this to work:

sales_dashboard/sales_dashboard$ redis-server (path where running this
from doesn't matter)

sales_dashboard/sales_dashboard$ celery -A sales_dashboard worker
--loglevel=DEBUG

sales_dashboard/sales_dashboard$ python3 manage.py runserver