/populate being managed by celery
The /populate view which gathers data from VTiger and updates
the database is now being managed by celery and therefore
the view functionality has been removed from views and added to
tasks. /populate still works so that this can be run manually.

Note, sometimes when changes are implemented, the celery worker
can't find them. In such a scenario, restart celery and clean
pycache.