/populate edits existing entries in the database
Prior to this commit, every time /populate was ran, there would be
new entries in the database. That means in the course of a day, there
could be hundreds of new rows. Instead, if there exists rows for today,
those rows will be simply updated instead of adding new ones.