Add date_modified field to models
Prior to this commit, date_created was being used to detect most
recent changes. The problem is that the date_created field
had auto_now_add=True and would only update once in the beginning
of the day. Any opportunity stages that were changed, would be added
to the main count and the same opportunity would cause a +1 increase
every time /populate was called. By checking against the date_modified
field, we only check the timeframe since /populate was last run and
not since the beginning of the day.