Restructuring VTigerAPI to disregard update times
The primary challenge was to account for the simple fact that the
same opportunity can change its stages multiple times per day.
The original plan was to collect all the data that occurred since
the time of the most recent update. This was causing some issues
and could be a problem if the server isn't running.

As a new solution to this issue, fields were added to the opportunity
for the times when the sales stage was changed. These fields are filled
out with the modified time via VTiger workflow:
https://i.imgur.com/sf4noGl.png
Therefore, instead of getting data from the past few minutes, we simply
overwrite the database with all the data gathered from the entire day.