Restructuing models.py to disregard update times
We are no longer gathering new results and adding them
to the database. Therefore, using "Sum" is not necessary.
The original intention was to have each '/populate'
generate new lines in the database every ten minutes
and then add them all together. Instead we are gathering
all data from VTiger and not needing to add anything together.