/populatecases URL now adds today's cases to db
A new URL has been added called "/populatecases" which
adds all the VTiger cases which have been modified today and
adds them into the Cases database. This is primarily a proof
of concept as this task will be offloaded to Celery and not
performed manually.
In addition, there isn't any duplicate checking in place yet
so when this URL is run again, all cases are duplicated.