Retrieve groups with IDs as they key
The goal here is to have the groups and usernames saved to a file
which can be referenced when pulling the cases so that API calls
to find the name of the user & group can be avoided.
If the username, or groupname can't be retrieved from the file, then
the file will be recreated.