Remove datepicker from Sales dashboard
This datepicker didn't do anything yet and was setup mainly to test.
The case dashboard date picker actually works and eventually the sales
date picker will be modeled the same way. Perhaps the two apps can
even share the same datepicker. For now, it will be removed so that
the auto refresh can be added to base.html and setup for both apps.