Fix calculation for resolved cases display
Resolved case counts which are displayed on the front end
now rely on the "case_resolved" field instead of the modified time. This
is because its possible for a case to be modified after
its been resolved. Especially since every resolved case
automatically becomes modified to closed after a certain amount
of days.