Removed End Date Selector - Date overhaul
Made a significant number of changes in this commit.
The cases that are presented no longer fit into a specified range.
Instead the user will select a single day and cases will be presented
for that specific day. If no date is selected, then it uses today's
current date as the default. The next step is to show the default
data for the current week and the current month.