Show all open cases on dashboard per group
Note: the DB must be populated with all the cases from the most
earliest "created time" or else this number won't be accurate.
This is because a group can have a case open from a year ago and
if that case is not in the db because its so old, then the amount
of total open cases will be minus 1.