Retrieve more than 100 modified cases per call
VTiger can only return a maximum of 100 cases per call. For situations
where there have been more than 100 modified cases in day, we now
return all of them. We need to use the "count" API call every time
in order to check this though. Still this is a critical step
to ensure we delete cases from the db which have been deleted in
vtiger.