Today's cases with outdated date_modified are deleted
A decent solution that works but only works for today's cases
that have been deleted.
If a case from yesterday gets deleted, then this method won't
catch it. Trying to avoid repleneshing the entire db, but any time
a case is deleted, the numbers won't match up. There's no way to find
out if a case has been deleted in VTiger, it simply vanishes from the
API.