Deleting old cases fixed. if amounts don't match up, repopulate db
The best all around solution to ensure that the numbers in the
dashboard match up to vtiger. The way to save all the API calls
would be to remove the ability of users to delete cases in VTiger
altogether and instead trigger a workflow to assign "deleted" cases
to a group nobody has permission to access.