Modifying "retrieve todays cases" for phonecalls/potentials
This begins the process of refactoring the sales side of the
dashboard to have two models of opportunities and phone calls
and do all the calculations from there. This will be faster
provide us with more data, allow us to look at historical data
without using any more api calls and reduce the overall number of
api calls.