Combine modified, created, resolved - recalculation
Displaying only the modified cases can be a problem.
The modifiedtime is the most recently modified time and will not
show if a case was modified during a time frame, it will only show
when it was most recently modified. Some cases can be resolved on
December 1st, but the modified time can be December 8th since
7 days after it was resolved, it gets closed and the modified time
updates. Therefore, the displayed cases need to match modified,
resolved and created. It may be worth while to remove the displayed
modified cases since its not an accurate number of when the case was
worked on.