Red & Green Resolved/Created Case rows
A row with a created case has a green background while a row with a
resolved case has a red background. This makes it easier to see at
a glance which cases contribute to the counts. If a case is both
created and resolved in the same time frame, it will be red. This
is because resolved cases has a higher priority.