WIP - User-Based Case Statistics
This is the beginning testing of generating user-based statistics
based on the data found in cases. Thus far, the data in /casestest
generates a list of all the cases assigned to the user along
with the average time spent of all the cases.
This can also be generated for specific time frames.