Collect first and last days of all months
The intention here is to loop through all the cases
in a given year and procure all the data for each month.

We can then showcase data from this month as it compares to last
month. Show the month's with the best kill ratio and the most
closed cases etc.