Data is generated for every month in every year
Instead of only having data for 12 months in the current year,
we generate data for every month in every year for every unique
year in the DB from all cases.
If a case exists in 2019, and a case exists in 2021,
then we'll have 36 sub-dicts. One for each 12 months in each 3
years.